Renamed a file as an excuse to write a note here:
There are unnecessary variables which I created for my convenience when still figuring out the coordinates, but which could now be removed to simplify the overall code.
However, it's late and I'm behind schedule as it is, and the code executes correctly, so I'll chalk it up to experience and leave it there.